apple-bom: rework paths traversal
The previous logic for resolving BomPath was incomplete in that it
failed to work if there were path references across multiple
BomBlockPaths instances.

This commit moves that logic to BomBlockTree, which is aware of the
existence of multiple paths blocks.

A few other minor tweaks were made as part of this commit to shore
up parsing and our understanding of how the data structure works.