apple-sdk: implement SDK searching
This isn't feature complete but it is a start. The commit already got
a bit away from me and I want to commit before it grows even larger.

As part of this, we invent a type to represent SDK version strings and
we change the AppleSdk trait and implementations to use this type
when appropriate. This type allows us to compare version strings
and sort SDKs by version number.