cpio-archive: restore deleted project
The content of the `cpio-archive` directory was restored as it last
existed in 125aac2f0061247b3d29447e2746ec01317d0ced, before it was
deleted in the commit immediately following.

Upon closer examination, `cpio-archive` is only used by
`apple-flat-package`, so I think it is fine for this crate to exist
in this repo.

The crate does not build and isn't part of the workspace. Fixups will
be performed in the next commit.