apple-codesign: more robust designated requirement handling
Fixes support for ADHOC app bundle signing with frameworks.

Before this, codesign would error with "the sealed resource directory
is invalid".

This add a new codepath when no code requirements has been set that
will enforce a requirement presence and ensure that appropriate
cdhashes are present for the given file.

Closes #44.