apple-codesign: match on parsed Subcommands
Now that we have the enum fully implemented, we can start constructing
it and matching on its variants. This gets us closer to passing the
argument structs into the command functions.

As part of this we hide the notarize commands when the feature isn't
enabled. Note that command hiding is inconsistent: sometimes we have
a fast failing command implementation variant sometimes we hide. We
can clean this up... later.