apple-codesign: add functionality for creating Mach-O files
The new `macho_builder` module contains functionality for creating a
Mach-O binary where none existed before. The goal of this code is to
facilitate testing of Mach-O signing.

Functionality is pretty basic for the moment. We don't even support
adding data to Mach-O sections / segments! But it's a start.

A `debug-create-macho` command has been added to expose this
functionality to the command line so it can be used by our trycmd
tests. This will eventually enable us to write trycmd tests for
signing.