apple-codesign: tungstenite 0.19 -> 0.20
As part of the upgrade several APIs were deprecated, including one
that had the side-effect of forcing flushes after a send. We port
to the new APIs. Rather than set ws settings to force a flush after
a send, we explicitly call flush() after sending a message so we're
explicit about behavior.