apple-codesign: shore up verify command
I saw a few issues where people attempted to run verify against
bundles, zips, and other files and ran into opaque errors. Let's
validate the path is a Mach-O before attempting to parse it as such

Also, since the verify command is known to be buggy, let's print a big
warning message front and center when it is used.