workspace: upgrade x509-certificate and cryptographic-message-syntax
This entailed a host of API breaks. Fortunately, I think it is pretty
straightforward.

The main thing we lack test coverage for is signing with RSA keys.
I plan to add that in a future commit. (The recent x509-certificate
changes were in support of making self-signed RSA certificates more
easily constructed.)