apple-codesign: rework code requirements derivation
Apple's heuristics appear to always derive a code requirements expression,
even for non Apple signed entities. In addition, it looks at the presence
of a CA extension to determine which expression to use.

We add support for deriving a non-Apple signed expression. And we rework
the existing logic to key off the CA extensions instead of certificate
profile.

Closes #99.