apple-codesign: add new Apple CA certificates
These were published since I last looked at the list of certs.

It looks like the new Apple Application Integration CA 7 introduced a
new extension. I'm assuming it has the same meaning as the extension on
the last variant of this CA as I couldn't find any documentation about this
OID anywhere on the Internet. But I didn't read the CPS PDFs in full.