apple-codesign: configuration files support
This took me far too long to implement. But I think it is in a
sufficiently mature state to merge.

Only the `sign` command has config settings so far.

Testing coverage isn't as robust as I would like. Specifically, I'm
a little concerned that config setting inheritance / merging is
broken in subtle ways. It took me a bit to figure out that `None`
can serialize and effectively overwrite subsequent values. So it
was important to omit empty struct fields, especially when that field
is itself a struct.

Closes #107.