apple-codesign: implement launch constraints and library constraints
These appear to share the same blob type and consist of a DER encoded
plist. There are 4 new special slots for the blobs.

The user-provided plist is supplemented with 3 additional fields and
the DER encoded form is a dict of those 4 fields. We don't yet know
what the other 3 fields are but they always appear to have the same
constant integer values. So we just hardcode things for now.

Closes #83.