apple-codesign: fold session initialization arguments into remote signing ones
The multiple levels of nesting combined with Option<T> were conspiring
to prevent CLI arguments from getting processed. I believe this is
https://github.com/clap-rs/clap/issues/4697 but am not 100% sure.

We mitigate the problem by moving the session initialization fields
into the main key struct so we have a single level of argument nesting.

As part of this we remote `--remote-signer` and instead infer remote
signing from presence of a session initialization argument.

We also make the server URL optional to prevent the struct from having
any populated fields by default.