apfs-types: new crate defining Apple file system core types and constants
This is the first step in implementing a pure Rust APFS library. I
believe we've implemented all APFS data structures and constants as
defined by Apple's documentation. But it is a 181 page specification
and I may have missed something.

This also introduces the apfs-derive crate implementing an
`derive(ApfsData)` macro used for generating parsers of APFS structs.