apple-codesign: exclude .dSYM/ directories from signing
This change is pretty straightforward. But it is a quick and dirty
workaround short of the proper solution of using the CodeResources
rules for determining which directories to recurse into. I don't
feel like incurring the effort to rewrite the bundle traversing logic
at this time. But it really needs to be done to finish closing all
these bundle traversal bugs.

Closes #128.