apple-codesign: add a shallow signing mode
We want to enable a signing mode that is similar to Apple's tooling in
that it doesn't traverse nested entities by default. This commit is the
start of that feature.

This should give people with complex signing requirements a way to opt
into multiple `sign` invocations instead of having to pass a series
of complex CLI arguments to control nested signing.