apple-codesign: prevent double signing / installing files when traversing bundles
Previously, it was possible to encounter a file multiple times when
signing a bundle containing child bundles. Normally we ignore traversing
into an already signed bundle. But this only works if the bundle matches
a CodeResources rule having the "nested" flag set. If a child bundle
were outside a "nested" directory, we would walk into it just like any
other directory.

It is important for us to walk into these non-"nested" child bundles
because the file content needs to be sealed - as files, not a bundle.

But this posed a problem with Mach-O binaries. If we signed a Mach-O
binary as part of a child bundle and then encountered that Mach-O binary
again as part of walking a non-"nested" child bundle, we would sign the
Mach-O again! In many scenarios this would break the child bundle's
signature since the Mach-O content changed.

And for non-Macho-O files, we may copy/write the destination file twice,
adding overhead to signing.

This commit introduces tracking of which files are installed when
signing a bundle. When signing a bundle, we track if a path has already
been installed. If so, we seal the already-written file (not the source
file!) and skip installing the file again.

This generic approach seems to magically fix issues with non-"nested"
child bundles.

The test changes clearly show that we stop double signing Mach-O binaries
located in non-"nested" child bundles. In the case of the child bundle's
main binary, the new signature correctly contains an Info.plist and
CodeResources file digests. This is because the original signed binary
is preserved. We also see that the tree of installed files is unchanged
modulo the a) now single-signed Mach-O, b) the CodeResources of the parent
bundle, and c) the parent bundle binary's signature. This all makes
sense.

Closes #149.